started changing our setup from JS to TS
