+++ backend/package.json
@@ -1,19 +1,20 @@
 {
   "name": "backend",
+  "type": "commonjs",
   "version": "1.0.0",
+  "description": "",
   "main": "index.js",
   "directories": {
     "test": "tests"
   },
   "scripts": {
-    "start": "node src/server.js",
-    "dev": "nodemon src/server.js",
+    "start": "node src/server.ts",
+    "dev": "nodemon src/server.ts",
     "test": "jest"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "description": "",
   "dependencies": {
     "cors": "^2.8.6",
     "dotenv": "^17.4.2",
@@ -21,7 +22,10 @@
     "pg": "^8.20.0"
   },
   "devDependencies": {
-    "jest": "^30.3.0",
-    "nodemon": "^3.1.14"
+    "@types/cors": "^2.8.19",
+    "@types/express": "^5.0.6",
+    "@types/node": "^25.7.0",
+    "ts-node": "^10.9.2",
+    "typescript": "^6.0.3"
   }
 }
