+++ docs/architecture diagram.drawio
