+++ frontend/vite.config.ts
@@ -1,12 +1,11 @@
 import { defineConfig } from 'vitest/config'
 import react from '@vitejs/plugin-react'
 
-
 export default defineConfig({
   plugins: [react()],
   test: {
     globals: true,
     environment: 'jsdom',
-    setupFiles: './testing/setup.js'
+    setupFiles: './test/setup.ts'
   },
 })
